Use long long for Species ID to prevent overflow errors

+++ include/MIMESim/Species.hpp
@@ -39,7 +39,7 @@ namespace species
     {
 
       private:
-        const unsigned int specId;
+        const unsigned long long specId;
         const constants::Constants& params;
         unsigned int count;
         const unsigned int numMut;
@@ -65,9 +65,9 @@ namespace species
         mutVector specIdxToMutPos();
 
       public:
-        Species(const unsigned int id, const constants::Constants& param);
+        Species(const unsigned long long id, const constants::Constants& param);
 
-        const unsigned int getSpecId() const;
+        const unsigned long long getSpecId() const;
 
         const constants::Constants& getParams() const;
 
@@ -124,15 +124,15 @@ namespace species
         double getFractionUnbound();
     };
 
-    using species_map = std::unordered_map<int, Species>;
+    using species_map = std::unordered_map<unsigned long long, Species>;
 
     species_map drawSpeciesIds(const constants::Constants& params);
 
-    mutVector specIdxToMutPos(const unsigned specId, const constants::Constants& params);
+    mutVector specIdxToMutPos(const unsigned long long specId, const constants::Constants& params);
 
-    unsigned mutPosToSpecIdx(const mutVector& mutPos, const constants::Constants& params);
+    unsigned long long mutPosToSpecIdx(const mutVector& mutPos, const constants::Constants& params);
 
-    unsigned getNumberOfMutationsById(const unsigned specId, const constants::Constants& params);
+    unsigned getNumberOfMutationsById(const unsigned long long specId, const constants::Constants& params);
 
     species_map readFromFile(const std::string& inputPath, utils::SampleID id, const constants::Constants& params);
